+++ arviz/plots/backends/bokeh/densityplot.py
@@ -6,9 +6,8 @@ import matplotlib.pyplot as plt
 import numpy as np
 from bokeh.models.annotations import Legend, Title
 
-from ....kde_utils import _kde
-from ....numeric_utils import get_bins, histogram
 from ....stats import hdi
+from ....stats.density_utils import get_bins, histogram, kde
 from ...plot_utils import _scale_fig_size, calculate_point_estimate, make_label, vectorized_to_hex
 from .. import show_layout
 from . import backend_kwarg_defaults, create_axes_grid
@@ -141,7 +140,7 @@ def _d_helper(
         else:
             new_vec = vec
 
-        x, density = _kde(new_vec, circular=circular, bw=bw)
+        x, density = kde(new_vec, circular=circular, bw=bw)
         density *= hdi_prob
         xmin, xmax = x[0], x[-1]
         ymin, ymax = density[0], density[-1]
