+++ arviz/plots/backends/bokeh/loopitplot.py
@@ -4,7 +4,7 @@ from bokeh.models import BoxAnnotation
 from matplotlib.colors import hsv_to_rgb, rgb_to_hsv, to_hex, to_rgb
 from xarray import DataArray
 
-from ....kde_utils import _kde
+from ....stats.density_utils import kde
 from ...plot_utils import _scale_fig_size
 from .. import show_layout
 from . import backend_kwarg_defaults, create_axes_grid
@@ -181,7 +181,7 @@ def plot_loo_pit(
             )
         else:
             for idx in range(n_unif):
-                x_s, unif_density = _kde(unif[idx, :])
+                x_s, unif_density = kde(unif[idx, :])
                 ax.line(
                     x_s,
                     unif_density,
