+++ arviz/plots/backends/bokeh/rankplot.py
@@ -5,7 +5,7 @@ from bokeh.models import Span
 from bokeh.models.annotations import Title
 from bokeh.models.tickers import FixedTicker
 
-from ....stats.stats_utils import histogram
+from ....stats.density_utils import histogram
 from ...plot_utils import _scale_fig_size, make_label
 from .. import show_layout
 from . import backend_kwarg_defaults, create_axes_grid
