+++ arviz/plots/backends/bokeh/violinplot.py
@@ -2,9 +2,8 @@
 import numpy as np
 from bokeh.models.annotations import Title
 
-from ....kde_utils import _kde
-from ....numeric_utils import get_bins, histogram
 from ....stats import hdi
+from ....stats.density_utils import get_bins, histogram, kde
 from ...plot_utils import _scale_fig_size, make_label
 from .. import show_layout
 from . import backend_kwarg_defaults, create_axes_grid
@@ -106,7 +105,7 @@ def _violinplot(val, rug, shade, bw, circular, ax, **shade_kwargs):
             bw = "taylor"
         else:
             bw = "experimental"
-    x, density = _kde(val, circular=circular, bw=bw)
+    x, density = kde(val, circular=circular, bw=bw)
 
     if not rug:
         x = np.concatenate([x, x[::-1]])
