+++ arviz/plots/backends/matplotlib/bpvplot.py
@@ -3,7 +3,7 @@ import matplotlib.pyplot as plt
 import numpy as np
 from scipy import stats
 
-from ....kde_utils import _kde
+from ....stats.density_utils import kde
 from ...kdeplot import plot_kde
 from ...plot_utils import (
     _scale_fig_size,
@@ -85,7 +85,7 @@ def plot_bpv(
 
         if kind == "p_value":
             tstat_pit = np.mean(pp_vals <= obs_vals, axis=-1)
-            x_s, tstat_pit_dens = _kde(tstat_pit)
+            x_s, tstat_pit_dens = kde(tstat_pit)
             ax_i.plot(x_s, tstat_pit_dens, linewidth=linewidth, color=color)
             ax_i.set_yticks([])
             if reference is not None:
@@ -104,7 +104,7 @@ def plot_bpv(
 
         elif kind == "u_value":
             tstat_pit = np.mean(pp_vals <= obs_vals, axis=0)
-            x_s, tstat_pit_dens = _kde(tstat_pit)
+            x_s, tstat_pit_dens = kde(tstat_pit)
             ax_i.plot(x_s, tstat_pit_dens, color=color)
             if reference is not None:
                 if reference == "analytical":
