+++ arviz/plots/backends/matplotlib/distplot.py
@@ -2,7 +2,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-from ....numeric_utils import get_bins
+from ....stats.density_utils import get_bins
 from ...kdeplot import plot_kde
 from ...plot_utils import _scale_fig_size
 from . import backend_kwarg_defaults, backend_show, create_axes_grid, matplotlib_kwarg_dealiaser
