+++ arviz/plots/loopitplot.py
@@ -2,9 +2,9 @@
 import numpy as np
 import scipy.stats as stats
 
-from ..kde_utils import _kde
 from ..rcparams import rcParams
 from ..stats import loo_pit as _loo_pit
+from ..stats.density_utils import kde
 from .plot_utils import get_plotting_function
 
 
@@ -160,7 +160,7 @@ def plot_loo_pit(
         )
         unif_ecdf = unif_ecdf / n_data_points
     else:
-        x_vals, loo_pit_kde = _kde(loo_pit)
+        x_vals, loo_pit_kde = kde(loo_pit)
 
         unif = np.random.uniform(size=(n_unif, loo_pit.size))
         if use_hdi:
