+++ arviz/stats/density_utils.py
@@ -1,15 +1,44 @@
-# pylint: disable=invalid-name
-"""Kernel density estimation functions for ArviZ."""
-
+# pylint: disable=invalid-name,too-many-lines
+"""Density estimation functions for ArviZ."""
 import warnings
 
 import numpy as np
 from scipy.fftpack import fft
 from scipy.optimize import brentq
-from scipy.signal import convolve, gaussian
+from scipy.signal import convolve, convolve2d, gaussian  # pylint: disable=no-name-in-module
+from scipy.sparse import coo_matrix
 from scipy.special import ive  # pylint: disable=no-name-in-module
 
-from .numeric_utils import histogram
+from ..utils import _cov, _dot, _stack, conditional_jit
+
+__all__ = ["kde", "histogram"]
+
+
+@conditional_jit(cache=True, nopython=True)
+def histogram(data, bins, range_hist=None):
+    """Conditionally jitted histogram.
+
+    Parameters
+    ----------
+    data : array-like
+        Input data. Passed as first positional argument to ``np.histogram``.
+    bins : int or array-like
+        Passed as keyword argument ``bins`` to ``np.histogram``.
+    range_hist : (float, float), optional
+        Passed as keyword argument ``range`` to ``np.histogram``.
+
+    Returns
+    -------
+    hist : array
+        The number of counts per bin.
+    density : array
+        The density corresponding to each bin.
+    bin_edges : array
+        The edges of the bins used.
+    """
+    hist, bin_edges = np.histogram(data, bins=bins, range=range_hist)
+    hist_dens = hist / (hist.sum() * np.diff(bin_edges))
+    return hist, hist_dens, bin_edges
 
 
 def _bw_scott(x, x_std=None, **kwargs):  # pylint: disable=unused-argument
@@ -192,8 +221,6 @@ def _a1inv(x):
 
 
 def _kappa_mle(x):
-    from .numeric_utils import _circular_mean
-
     mean = _circular_mean(x)
     kappa = _a1inv(np.mean(np.cos(x - mean)))
     return kappa
@@ -225,6 +252,7 @@ def _dct1d(x):
     return output
 
 
+@conditional_jit(cache=True, nopython=True)
 def _fixed_point(t, N, k_sq, a_sq):
     """Calculate t-zeta*gamma^[l](t).
 
@@ -243,7 +271,7 @@ def _fixed_point(t, N, k_sq, a_sq):
     f = np.sum(np.power(k_sq, l) * a_sq * np.exp(-k_sq * np.pi ** 2 * t))
     f *= 0.5 * np.pi ** (2.0 * l)
 
-    for j in reversed(range(2, l)):
+    for j in np.arange(l - 1, 2 - 1, -1):
         c1 = (1 + 0.5 ** (j + 0.5)) / 3
         c2 = np.product(np.arange(1.0, 2 * j + 1, 2, dtype=np.float64))
         c2 /= (np.pi / 2) ** 0.5
@@ -255,10 +283,11 @@ def _fixed_point(t, N, k_sq, a_sq):
     return out
 
 
+@conditional_jit(cache=True, nopython=True)
 def _root(function, N, args, x):
     # The right bound is at most 0.01
     found = False
-    N = max(min(1050, N), 50)
+    N = np.max(np.min(1050, N), 50)
     tol = 10e-12 + 0.01 * (N - 50) / 1000
 
     while not found:
@@ -421,7 +450,7 @@ def _get_grid(
     return grid_min, grid_max, grid_len
 
 
-def _kde(x, circular=False, **kwargs):
+def kde(x, circular=False, **kwargs):
     """One dimensional density estimation.
 
     It is a wrapper around `kde_linear()` and `kde_circular()`.
@@ -451,10 +480,10 @@ def _kde(x, circular=False, **kwargs):
 
     >>> import numpy as np
     >>> import matplotlib.pyplot as plt
-    >>> from arviz.kde_utils import _kde
+    >>> from arviz import kde
     >>>
     >>> rvs = np.random.gamma(shape=1.8, size=1000)
-    >>> grid, pdf = _kde(rvs)
+    >>> grid, pdf = kde(rvs)
     >>> plt.plot(grid, pdf)
     >>> plt.show()
 
@@ -462,7 +491,7 @@ def _kde(x, circular=False, **kwargs):
     .. plot::
         :context: close-figs
 
-    >>> grid, pdf = _kde(rvs, bw="silverman")
+    >>> grid, pdf = kde(rvs, bw="silverman")
     >>> plt.plot(grid, pdf)
     >>> plt.show()
 
@@ -471,7 +500,7 @@ def _kde(x, circular=False, **kwargs):
         :context: close-figs
 
     >>> # bw_fct > 1 means more smoothness.
-    >>> grid, pdf = _kde(rvs, bw_fct=2.5)
+    >>> grid, pdf = kde(rvs, bw_fct=2.5)
     >>> plt.plot(grid, pdf)
     >>> plt.show()
 
@@ -479,7 +508,7 @@ def _kde(x, circular=False, **kwargs):
     .. plot::
         :context: close-figs
 
-    >>> grid, pdf = _kde(rvs, bound_correction=False, extend=True, extend_fct=0.5)
+    >>> grid, pdf = kde(rvs, bound_correction=False, extend=True, extend_fct=0.5)
     >>> plt.plot(grid, pdf)
     >>> plt.show()
 
@@ -487,7 +516,7 @@ def _kde(x, circular=False, **kwargs):
     .. plot::
         :context: close-figs
     # It accepts tuples and lists of length 2.
-    >>> grid, pdf = __kde(rvs, bound_correction=False, custom_lims=(0, 10))
+    >>> grid, pdf = kde(rvs, bound_correction=False, custom_lims=(0, 10))
     >>> plt.plot(grid, pdf)
     >>> plt.show()
 
@@ -496,7 +525,7 @@ def _kde(x, circular=False, **kwargs):
         :context: close-figs
 
     >>> rvs = np.random.vonmises(mu=np.pi, kappa=1, size=500)
-    >>> grid, pdf = _kde(rvs, circular=True)
+    >>> grid, pdf = kde(rvs, circular=True)
     >>> plt.plot(grid, pdf)
     >>> plt.show()
 
@@ -506,7 +535,7 @@ def _kde(x, circular=False, **kwargs):
 
     >>> rvs = np.random.vonmises(mu=np.pi, kappa=1, size=500)
     >>> # bw_fct > 1 means less smoothness.
-    >>> grid, pdf = _kde(rvs, circular=True, bw_fct=3)
+    >>> grid, pdf = kde(rvs, circular=True, bw_fct=3)
     >>> plt.plot(grid, pdf)
     >>> plt.show()
 
@@ -515,14 +544,14 @@ def _kde(x, circular=False, **kwargs):
         :context: close-figs
     >>> # This is still experimental, does not always work.
     >>> rvs = np.random.vonmises(mu=0, kappa=30, size=500)
-    >>> grid, pdf = _kde(rvs, circular=True, custom_lims=(-1, 1))
+    >>> grid, pdf = kde(rvs, circular=True, custom_lims=(-1, 1))
     >>> plt.plot(grid, pdf)
     >>> plt.show()
 
     See Also
     --------
     plot_kde : Compute and plot a kernel density estimate.
-    arviz.kde_utils._kde: Arviz KDE estimator
+    arviz.stats.density_utils.kde: Arviz KDE estimator
     """
     if circular:
         kde_fun = _kde_circular
@@ -690,8 +719,6 @@ def _kde_circular(
         return np.array([np.nan]), np.array([np.nan])
 
     # All values between -pi and pi
-    from .numeric_utils import _normalize_angle
-
     x = _normalize_angle(x)
 
     # Check `bw_fct` is numeric and positive
@@ -751,7 +778,7 @@ def _kde_convolution(x, bw, grid_edges, grid_counts, grid_len, bound_correction)
 
     One dimensional Gaussian kernel density estimation via
     convolution of the binned relative frequencies and a Gaussian filter.
-    This is an internal function used by `_kde()`.
+    This is an internal function used by `kde()`.
     """
     # Calculate relative frequencies per bin
     bin_width = grid_edges[1] - grid_edges[0]
@@ -788,7 +815,7 @@ def _kde_adaptive(x, bw, grid_edges, grid_counts, grid_len, bound_correction):
     The implementation uses the binning technique.
     Since there is not an unique `bw`, the convolution is not possible.
     The alternative implemented in this function is known as Abramson's method.
-    This is an internal function used by `_kde()`.
+    This is an internal function used by `kde()`.
     """
     # Pilot computations used for bandwidth adjustment
     pilot_grid, pilot_pdf = _kde_convolution(
@@ -840,3 +867,177 @@ def _kde_adaptive(x, bw, grid_edges, grid_counts, grid_len, bound_correction):
         pdf = np.sum(pdf_mat, axis=0) / len(x)
 
     return grid, pdf
+
+
+def _fast_kde(x, cumulative=False, bw=4.5, xmin=None, xmax=None):  # pylint: disable=unused-argument
+    """Kernel Density Estimate, Deprecated."""
+    if not (xmin is None and xmax is None):
+        custom_lims = (xmin, xmax)
+    else:
+        custom_lims = None
+    grid, pdf = kde(x, cumulative=cumulative, bw=bw, custom_lims=custom_lims)
+
+    warnings.warn("_fast_kde() has been replaced by kde() in stats.density_utils.py", FutureWarning)
+    return grid, pdf
+
+
+def _fast_kde_2d(x, y, gridsize=(128, 128), circular=False):
+    """
+    2D fft-based Gaussian kernel density estimate (KDE).
+
+    The code was adapted from https://github.com/mfouesneau/faststats
+
+    Parameters
+    ----------
+    x : Numpy array or list
+    y : Numpy array or list
+    gridsize : tuple
+        Number of points used to discretize data. Use powers of 2 for fft optimization
+    circular: bool
+        If True, use circular boundaries. Defaults to False
+    Returns
+    -------
+    grid: A gridded 2D KDE of the input points (x, y)
+    xmin: minimum value of x
+    xmax: maximum value of x
+    ymin: minimum value of y
+    ymax: maximum value of y
+    """
+    x = np.asarray(x, dtype=float)
+    x = x[np.isfinite(x)]
+    y = np.asarray(y, dtype=float)
+    y = y[np.isfinite(y)]
+
+    xmin, xmax = x.min(), x.max()
+    ymin, ymax = y.min(), y.max()
+
+    len_x = len(x)
+    weights = np.ones(len_x)
+    n_x, n_y = gridsize
+
+    d_x = (xmax - xmin) / (n_x - 1)
+    d_y = (ymax - ymin) / (n_y - 1)
+
+    xyi = _stack(x, y).T
+    xyi -= [xmin, ymin]
+    xyi /= [d_x, d_y]
+    xyi = np.floor(xyi, xyi).T
+
+    scotts_factor = len_x ** (-1 / 6)
+    cov = _cov(xyi)
+    std_devs = np.diag(cov) ** 0.5
+    kern_nx, kern_ny = np.round(scotts_factor * 2 * np.pi * std_devs)
+
+    inv_cov = np.linalg.inv(cov * scotts_factor ** 2)
+
+    x_x = np.arange(kern_nx) - kern_nx / 2
+    y_y = np.arange(kern_ny) - kern_ny / 2
+    x_x, y_y = np.meshgrid(x_x, y_y)
+
+    kernel = _stack(x_x.flatten(), y_y.flatten())
+    kernel = _dot(inv_cov, kernel) * kernel
+    kernel = np.exp(-kernel.sum(axis=0) / 2)
+    kernel = kernel.reshape((int(kern_ny), int(kern_nx)))
+
+    boundary = "wrap" if circular else "symm"
+
+    grid = coo_matrix((weights, xyi), shape=(n_x, n_y)).toarray()
+    grid = convolve2d(grid, kernel, mode="same", boundary=boundary)
+
+    norm_factor = np.linalg.det(2 * np.pi * cov * scotts_factor ** 2)
+    norm_factor = len_x * d_x * d_y * norm_factor ** 0.5
+
+    grid /= norm_factor
+
+    return grid, xmin, xmax, ymin, ymax
+
+
+def get_bins(values):
+    """
+    Automatically compute the number of bins for discrete variables.
+
+    Parameters
+    ----------
+    values = numpy array
+        values
+
+    Returns
+    -------
+    array with the bins
+
+    Notes
+    -----
+    Computes the width of the bins by taking the maximun of the Sturges and the Freedman-Diaconis
+    estimators. According to numpy `np.histogram` this provides good all around performance.
+
+    The Sturges is a very simplistic estimator based on the assumption of normality of the data.
+    This estimator has poor performance for non-normal data, which becomes especially obvious for
+    large data sets. The estimate depends only on size of the data.
+
+    The Freedman-Diaconis rule uses interquartile range (IQR) to estimate the binwidth.
+    It is considered a robusts version of the Scott rule as the IQR is less affected by outliers
+    than the standard deviation. However, the IQR depends on fewer points than the standard
+    deviation, so it is less accurate, especially for long tailed distributions.
+    """
+    x_min = values.min().astype(int)
+    x_max = values.max().astype(int)
+
+    # Sturges histogram bin estimator
+    bins_sturges = (x_max - x_min) / (np.log2(values.size) + 1)
+
+    # The Freedman-Diaconis histogram bin estimator.
+    iqr = np.subtract(*np.percentile(values, [75, 25]))  # pylint: disable=assignment-from-no-return
+    bins_fd = 2 * iqr * values.size ** (-1 / 3)
+
+    width = np.round(np.max([1, bins_sturges, bins_fd])).astype(int)
+
+    return np.arange(x_min, x_max + width + 1, width)
+
+
+def _sturges_formula(dataset, mult=1):
+    """Use Sturges' formula to determine number of bins.
+
+    See https://en.wikipedia.org/wiki/Histogram#Sturges'_formula
+    or https://doi.org/10.1080%2F01621459.1926.10502161
+
+    Parameters
+    ----------
+    dataset: xarray.DataSet
+       Must have the `draw` dimension
+
+    mult: float
+        Used to scale the number of bins up or down. Default is 1 for Sturges' formula.
+
+    Returns
+    -------
+    int
+        Number of bins to use
+    """
+    return int(np.ceil(mult * np.log2(dataset.draw.size)) + 1)
+
+
+def _circular_mean(x, na_rm=False):
+    """Compute mean of circular variable measured in radians.
+
+    The result is between -pi and pi.
+    """
+    if na_rm:
+        x = x[~np.isnan(x)]
+
+    sinr = np.sum(np.sin(x))
+    cosr = np.sum(np.cos(x))
+    mean = np.arctan2(sinr, cosr)
+
+    return mean
+
+
+def _normalize_angle(x, zero_centered=True):
+    """Normalize angles.
+
+    Take angles in radians and normalize them to [-pi, pi) or [0, 2 * pi)
+    depending on `zero_centered`.
+    """
+    if zero_centered:
+        return (x + np.pi) % (2 * np.pi) - np.pi
+    else:
+        return x % (2 * np.pi)
