+++ arviz/stats/stats.py
@@ -13,10 +13,11 @@ import xarray as xr
 from scipy.optimize import minimize
 
 from ..data import CoordSpec, DimSpec, InferenceData, convert_to_dataset, convert_to_inference_data
-from ..kde_utils import _kde
-from ..numeric_utils import get_bins, histogram
 from ..rcparams import rcParams
 from ..utils import Numba, _numba_var, _var_names, credible_interval_warning, get_coords
+from .density_utils import get_bins as _get_bins
+from .density_utils import histogram as _histogram
+from .density_utils import kde as _kde
 from .diagnostics import _mc_error, _multichain_statistics, ess
 from .stats_utils import ELPDData, _circular_standard_deviation
 from .stats_utils import get_log_likelihood as _get_log_likelihood
@@ -548,8 +549,8 @@ def _hdi_multimodal(ary, hdi_prob, skipna, max_modes):
         range_x = upper - lower
         dx = range_x / len(density)
     else:
-        bins = get_bins(ary)
-        _, density, _ = histogram(ary, bins=bins)
+        bins = _get_bins(ary)
+        _, density, _ = _histogram(ary, bins=bins)
         dx = np.diff(bins)[0]
 
     density *= dx
